Rename all "model" variables and modules with full names
Code is read many times more than it is written. Though it is quicker to write `lm` etc, it is unecessarily confusing for a first time reader, especially for a module name. Give full names to these concepts so they are clearer for a reader
