+++ allosaurus/app.py
@@ -1,9 +1,9 @@
-from allosaurus.am.utils import *
+from allosaurus.acoustic_model.utils import *
 from pathlib import Path
 from allosaurus.audio import read_audio
-from allosaurus.pm.factory import read_pm
-from allosaurus.am.factory import read_am
-from allosaurus.lm.factory import read_lm
+from allosaurus.preprocess_model.factory import read_preprocess_model
+from allosaurus.acoustic_model.factory import read_acoustic_model
+from allosaurus.language_model.factory import read_language_model
 from allosaurus.bin.download_model import download_model
 from allosaurus.model import resolve_model_name, get_all_models
 from argparse import Namespace
@@ -35,24 +35,24 @@ def read_recognizer(inference_config_or_name='latest', alt_model_path=None):
 
     assert model_path.exists(), f"{inference_config.model} is not a valid model"
 
-    # create pm (pm stands for preprocess model: audio -> feature etc..)
-    pm = read_pm(model_path, inference_config)
+    # preprocess(audio) -> features etc..
+    preprocess_model = read_preprocess_model(model_path, inference_config)
 
-    # create am (acoustic model: feature -> logits )
-    am = read_am(model_path, inference_config)
+    # acoustic(features) -> logits
+    acoustic_model = read_acoustic_model(model_path, inference_config)
 
-    # create lm (language model: logits -> phone)
-    lm = read_lm(model_path, inference_config)
+    # language(logits) -> phones
+    language_model = read_language_model(model_path, inference_config)
 
-    return Recognizer(pm, am, lm, inference_config)
+    return Recognizer(preprocess_model, acoustic_model, language_model, inference_config)
 
 class Recognizer:
 
-    def __init__(self, pm, am, lm, config):
+    def __init__(self, preprocess_model, acoustic_model, language_model, config):
 
-        self.pm = pm
-        self.am = am
-        self.lm = lm
+        self.preprocess_model = preprocess_model
+        self.acoustic_model = acoustic_model
+        self.language_model = language_model
         self.config = config
 
     def is_available(self, lang_id):
@@ -71,7 +71,7 @@ class Recognizer:
         audio = read_audio(filename)
 
         # extract feature
-        feat = self.pm.compute(audio)
+        feat = self.preprocess_model.compute(audio)
 
         # add batch dim
         feats = np.expand_dims(feat, 0)
@@ -79,12 +79,12 @@ class Recognizer:
 
         tensor_batch_feat, tensor_batch_feat_len = move_to_tensor([feats, feat_len], self.config.device_id)
 
-        tensor_batch_lprobs = self.am(tensor_batch_feat, tensor_batch_feat_len)
+        tensor_batch_lprobs = self.acoustic_model(tensor_batch_feat, tensor_batch_feat_len)
 
         if self.config.device_id >= 0:
             batch_lprobs = tensor_batch_lprobs.cpu().detach().numpy()
         else:
             batch_lprobs = tensor_batch_lprobs.detach().numpy()
 
-        token = self.lm.compute(batch_lprobs[0], lang_id, topk, emit=emit, timestamp=timestamp)
+        token = self.language_model.compute(batch_lprobs[0], lang_id, topk, emit=emit, timestamp=timestamp)
         return token
