+++ allosaurus/bin/__init__.py
@@ -1,0 +1,3 @@
+"""
+Bin module provides entrypoint scripts for various utility functions.
+"""
