+++ allosaurus/bin/list_lang.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 from allosaurus.model import get_model_path
-from allosaurus.lm.inventory import Inventory
+from allosaurus.language_model.inventory import Inventory
 import argparse
 
 if __name__ == '__main__':
