+++ allosaurus/bin/restore_phone.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-from allosaurus.lm.inventory import Inventory
+from allosaurus.language_model.inventory import Inventory
 from allosaurus.model import get_model_path
 import argparse
